src/frontend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
attrs==19.3.0
    # via
    #   -r requirements.in
    #   pytest
bcrypt==3.1.7
    # via -r requirements.in
bleach==3.3.0
    # via -r requirements.in
cachetools==4.1.1
    # via
    #   -r requirements.in
    #   google-auth
certifi==2020.6.20
    # via
    #   -r requirements.in
    #   requests
cffi==1.14.0
    # via
    #   -r requirements.in
    #   bcrypt
    #   cryptography
chardet==3.0.4
    # via
    #   -r requirements.in
    #   requests
click==7.1.1
    # via
    #   -r requirements.in
    #   flask
coverage==5.3
    # via pytest-cov
cryptography==3.3.2
    # via -r requirements.in
email_validator==1.1.3
flask==1.1.2
    # via
    #   -r requirements.in
    #   opentelemetry-instrumentation-flask
flask_wtf==1.0.1
google-api-core[grpc]==1.22.2
    # via
    #   -r requirements.in
    #   google-cloud-core
    #   google-cloud-monitoring
    #   google-cloud-trace
google-auth==1.21.2
    # via
    #   -r requirements.in
    #   google-api-core
    #   opentelemetry-tools-google-cloud
google-cloud-core==1.4.1
    # via
    #   -r requirements.in
    #   google-cloud-trace
google-cloud-monitoring==1.1.0
    # via opentelemetry-exporter-google-cloud
google-cloud-trace==0.24.0
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-google-cloud
googleapis-common-protos==1.52.0
    # via
    #   -r requirements.in
    #   google-api-core
grpcio==1.30.0
    # via
    #   -r requirements.in
    #   google-api-core
gunicorn==20.1.0
    # via -r requirements.in
idna==2.10
    # via
    #   -r requirements.in
    #   requests
importlib-metadata==1.7.0
    # via -r requirements.in
itsdangerous==1.1.0
    # via
    #   -r requirements.in
    #   flask
jinja2==2.11.3
    # via
    #   -r requirements.in
    #   flask
markupsafe==1.1.1
    # via
    #   -r requirements.in
    #   jinja2
more-itertools==8.4.0
    # via
    #   -r requirements.in
    #   pytest
opentelemetry-api==0.13b0
    # via
    #   opentelemetry-exporter-google-cloud
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-sqlalchemy
    #   opentelemetry-instrumentation-wsgi
    #   opentelemetry-sdk
    #   opentelemetry-tools-google-cloud
opentelemetry-exporter-google-cloud==0.13b0
    # via -r requirements.in
opentelemetry-instrumentation-flask==0.13b0
    # via -r requirements.in
opentelemetry-instrumentation-sqlalchemy==0.13b0
    # via -r requirements.in
opentelemetry-instrumentation-wsgi==0.13b0
    # via opentelemetry-instrumentation-flask
opentelemetry-instrumentation==0.13b0
    # via
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-sqlalchemy
    #   opentelemetry-instrumentation-wsgi
opentelemetry-sdk==0.13b0
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-google-cloud
    #   opentelemetry-tools-google-cloud
opentelemetry-tools-google-cloud==0.13b0
    # via -r requirements.in
packaging==20.4
    # via
    #   -r requirements.in
    #   bleach
    #   pytest
pluggy==0.13.1
    # via
    #   -r requirements.in
    #   pytest
protobuf==3.12.2
    # via
    #   -r requirements.in
    #   google-api-core
    #   googleapis-common-protos
psycopg2==2.7.7
    # via -r requirements.in
py==1.10.0
    # via
    #   -r requirements.in
    #   pytest
pyasn1-modules==0.2.8
    # via
    #   -r requirements.in
    #   google-auth
pyasn1==0.4.8
    # via
    #   -r requirements.in
    #   pyasn1-modules
    #   rsa
pycparser==2.20
    # via
    #   -r requirements.in
    #   cffi
pycryptodome==3.9.7
    # via -r requirements.in
pyjwt==1.7.1
    # via -r requirements.in
pyparsing==2.4.7
    # via
    #   -r requirements.in
    #   packaging
pytest-cov==2.10.0
    # via -r requirements.in
pytest==5.4.1
    # via
    #   -r requirements.in
    #   pytest-cov
pytz==2020.1
    # via
    #   -r requirements.in
    #   google-api-core
requests==2.24.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   opentelemetry-tools-google-cloud
rsa==4.7
    # via
    #   -r requirements.in
    #   google-auth
six==1.14.0
    # via
    #   -r requirements.in
    #   bcrypt
    #   bleach
    #   cryptography
    #   google-api-core
    #   google-auth
    #   grpcio
    #   packaging
    #   protobuf
sqlalchemy==1.3.16
    # via
    #   -r requirements.in
    #   opentelemetry-instrumentation-sqlalchemy
urllib3==1.25.9
    # via
    #   -r requirements.in
    #   requests
wcwidth==0.2.5
    # via
    #   -r requirements.in
    #   pytest
webencodings==0.5.1
    # via
    #   -r requirements.in
    #   bleach
werkzeug==1.0.1
    # via
    #   -r requirements.in
    #   flask
wrapt==1.12.1
    # via
    #   -r requirements.in
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-sqlalchemy
zipp==1.2.0
    # via
    #   -r requirements.in
    #   importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools